+++ src/main/java/frc/robot/commands/TrajectoryGen.java
@@ -1,3 +1,4 @@
+package frc.robot.commands;
 // package frc.robot.autonomous;
 
 // import java.util.Optional;
